-- Bug founded at step file - wrong driver
Using another webdriver with chrome won't work. In the course he useing FireFox and gecko.driver, but I have been using chrome. So to change that: I have updated my chrome, because it still works with the newest updat while you are using 32 bit driver. I used a other path: wedriver.chrome.driver and the path where this is. With this I was able to update and execute the before code and the first two @given steps